Implemented 2d stairs.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@941 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ heroquest/HQClient/hero.cpp
@@ -201,7 +201,7 @@ void Hero::createFromGui(Gui::GuiElement*){
 void Hero::write(const string& p) const{
   ofstream out(p.c_str(), ios::binary);
 
-  float version = VERSION;
+  float version = (float)VERSION;
   out.write((char*)&version, sizeof(version));
 
   short length = player_.length();
